+++ app/src/main/java/com/codepatissier/keki/src/main/consumer/mypage/profileEdit/ConsumerProfileEditRetrofitInterface.kt
@@ -1,0 +1,10 @@
+package com.codepatissier.keki.src.main.consumer.mypage.profileEdit
+
+import retrofit2.http.PATCH
+
+interface ConsumerProfileEditRetrofitInterface {
+    @PATCH("/users/profile")
+    fun editProfile(
+        @Path("")
+    ) : Call<>
+}
