+++ app/src/main/java/com/codepatissier/keki/src/main/consumer/mypage/profileEdit/model/ConsumerProfileEditResponse.kt
@@ -1,0 +1,7 @@
+package com.codepatissier.keki.src.main.consumer.mypage.profileEdit.model
+
+data class ConsumerProfileEditResponse (
+    val code: Int,
+    val isSuccess: Boolean,
+    val message: String
+)
