+++ app/src/main/res/layout/activity_consumer_profile_edit.xml
@@ -28,6 +28,7 @@
         app:layout_constraintStart_toEndOf="@id/iv_back"
         app:layout_constraintTop_toTopOf="parent"/>
     <TextView
+        android:id="@+id/tv_edit"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:includeFontPadding="false"
